Move sshjump analysis to kubernetes_utils.
